Rework communication system while playing
The server now sends current game state to every player every 30 ms (on current map). The player's moving now also happens in the client and after moving it informs server about its new location

+++ Server/Game/Units/Abstract/GenericUnit.cs
@@ -30,7 +30,7 @@ namespace Server
 
         public virtual void PlayCycle(int timeSpan)
         {
-            float movePoints = timeSpan / SLOWING_CONSTANT;
+            int movePoints = timeSpan / SLOWING_CONSTANT;
 
             if (Direction == MovingDirection.Right)
                 System.Console.WriteLine("TS: " + timeSpan + ", MP: " + movePoints);
