+++ Shared/Location.cs
@@ -3,23 +3,23 @@
     public class Location
     {
         public int MapID { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
 
-        public Location(int mapId, float x, float y)
+        public Location(int mapId, int x, int y)
         {
             this.MapID = mapId;
             this.X = x;
             this.Y = y;
         }
 
-        public Location(float x, float y) 
+        public Location(int x, int y) 
             :this(-1, x, y)
         {
         }
 
         public Location()
-            :this(-1, -1f, -1f)
+            :this(-1, -1, -1)
         {
         }
     }
